+++ spec/architecture.spec.ts
@@ -2,7 +2,7 @@ import * as architecture from '../src/architecture'
 import {Architecture} from '../src/architecture'
 import {Host} from '../src/host'
 import * as os from '../src/operating_systems/kind'
-import {Qemu, QemuEfi, Xhyve} from '../src/hypervisor'
+import {Qemu, QemuEfi} from '../src/hypervisor'
 
 let context = describe
 
@@ -10,36 +10,6 @@ describe('Architecture', () => {
   describe('efiHypervisor', () => {
     context('x86_64', () => {
       let kind = architecture.Kind.x86_64
-
-      context('macOS host', () => {
-        let host = Host.create('darwin')
-
-        context('QEMU hypervisor', () => {
-          let selectedHypervisor = new Qemu()
-
-          context('OpenBSD', () => {
-            let osKind = os.Kind.for('openbsd')
-            let arch = Architecture.for(kind, host, osKind, selectedHypervisor)
-
-            it('returns the QEMU EFI hypervisor', () => {
-              expect(arch.efiHypervisor).toBeInstanceOf(QemuEfi)
-            })
-          })
-        })
-
-        context('Xhyve hypervisor', () => {
-          let selectedHypervisor = new Xhyve()
-
-          context('OpenBSD', () => {
-            let osKind = os.Kind.for('openbsd')
-            let arch = Architecture.for(kind, host, osKind, selectedHypervisor)
-
-            it('returns the Xhyve hypervisor', () => {
-              expect(arch.efiHypervisor).toBeInstanceOf(Xhyve)
-            })
-          })
-        })
-      })
 
       context('Linux host', () => {
         let host = Host.create('linux')
@@ -63,36 +33,6 @@ describe('Architecture', () => {
   describe('hypervisor', () => {
     context('x86_64', () => {
       let kind = architecture.Kind.x86_64
-
-      context('macOS host', () => {
-        let host = Host.create('darwin')
-
-        context('QEMU hypervisor', () => {
-          let selectedHypervisor = new Qemu()
-
-          context('OpenBSD', () => {
-            let osKind = os.Kind.for('openbsd')
-            let arch = Architecture.for(kind, host, osKind, selectedHypervisor)
-
-            it('returns the QEMU EFI hypervisor', () => {
-              expect(arch.hypervisor).toBeInstanceOf(Qemu)
-            })
-          })
-        })
-
-        context('Xhyve hypervisor', () => {
-          let selectedHypervisor = new Xhyve()
-
-          context('OpenBSD', () => {
-            let osKind = os.Kind.for('openbsd')
-            let arch = Architecture.for(kind, host, osKind, selectedHypervisor)
-
-            it('returns the Xhyve hypervisor', () => {
-              expect(arch.hypervisor).toBeInstanceOf(Xhyve)
-            })
-          })
-        })
-      })
 
       context('Linux host', () => {
         let host = Host.create('linux')