+++ spec/operating_systems/netbsd/netbsd.spec.ts
@@ -5,13 +5,12 @@ import * as os from '../../../src/operating_systems/kind'
 import HostQemu from '../../../src/host_qemu'
 import * as hypervisor from '../../../src/hypervisor'
 import * as qemu from '../../../src/qemu_vm'
-import * as xhyve from '../../../src/xhyve_vm'
 import * as netbsdQemuVm from '../../../src/operating_systems/netbsd/qemu_vm'
 import {Input} from '../../../src/action/input'
 
 describe('NetBSD OperatingSystem', () => {
   class Host extends hostModule.Host {
-    get vmModule(): typeof xhyve | typeof qemu {
+    get vmModule(): typeof qemu {
       return qemu
     }
 
@@ -34,8 +33,6 @@ describe('NetBSD OperatingSystem', () => {
     override get defaultCpuCount(): number {
       return 6
     }
-
-    override validateHypervisor(_kind: hypervisor.Kind): void {}
   }
 
   let host = new Host()
@@ -50,14 +47,13 @@ describe('NetBSD OperatingSystem', () => {
   let hypervisorDirectory = 'hypervisor/directory'
   let resourcesDirectory = 'resources/directory'
   let firmwareDirectory = 'firmware/directory'
-  let input = new Input()
+  let input = new Input(host)
 
   let config = {
     memory: '4G',
     cpuCount: 7,
     diskImage: '',
-    resourcesDiskImage: '',
-    userboot: ''
+    resourcesDiskImage: ''
   }
 
   describe('createVirtualMachine', () => {
@@ -82,78 +78,9 @@ describe('NetBSD OperatingSystem', () => {
           ssHostPort: 2847,
           cpu: 'max',
           machineType: 'q35',
-          uuid: '864ED7F0-7876-4AA7-8511-816FABCFA87F',
           firmware: `${firmwareDirectory}/share/qemu/bios-256k.bin`
         }
       )
-    })
-
-    describe('when on a macOS host', () => {
-      class Host extends hostModule.Host {
-        get vmModule(): typeof xhyve | typeof qemu {
-          return xhyve
-        }
-
-        override get qemu(): HostQemu {
-          return new HostQemu.MacosHostQemu()
-        }
-
-        override get hypervisor(): hypervisor.Hypervisor {
-          return new hypervisor.Xhyve()
-        }
-
-        override get efiHypervisor(): hypervisor.Hypervisor {
-          return this.hypervisor
-        }
-
-        override get defaultMemory(): string {
-          return '13G'
-        }
-
-        override get defaultCpuCount(): number {
-          return 3
-        }
-
-        override validateHypervisor(_kind: hypervisor.Kind): void {}
-      }
-
-      let host = new Host()
-      let osKind = os.Kind.for('netbsd')
-      const vmm = new hypervisor.Qemu()
-      let architecture = arch.Architecture.for(
-        arch.Kind.x86_64,
-        host,
-        osKind,
-        vmm
-      )
-      let netbsd = new NetBsd(architecture, '0.0.0')
-
-      it('creates a virtual machine configured with BIOS firmware', () => {
-        let qemuVmSpy = spyOn(netbsdQemuVm, 'Vm')
-
-        netbsd.createVirtualMachine(
-          hypervisorDirectory,
-          resourcesDirectory,
-          firmwareDirectory,
-          input,
-          config
-        )
-
-        expect(qemuVmSpy).toHaveBeenCalledOnceWith(
-          hypervisorDirectory,
-          resourcesDirectory,
-          architecture,
-          input,
-          {
-            ...config,
-            ssHostPort: 2847,
-            cpu: 'max',
-            machineType: 'q35',
-            uuid: '864ED7F0-7876-4AA7-8511-816FABCFA87F',
-            firmware: `${firmwareDirectory}/share/qemu/bios-256k.bin`
-          }
-        )
-      })
     })
   })
 })