+++ spec/operating_systems/omnios/qemu_vm.spec.ts
@@ -1,6 +1,6 @@
 import {Vm} from '../../../src/operating_systems/omnios/qemu_vm'
 import * as arch from '../../../src/architecture'
-import {host} from '../../../src/host'
+import {Host} from '../../../src/host'
 import * as os from '../../../src/operating_systems/kind'
 import '../../../src/operating_systems/omnios/omnios'
 import {Input} from '../../../src/action/input'
@@ -11,6 +11,7 @@ describe('OmniOS QemuVm', () => {
   let cpuCount = 8
   let ssHostPort = 5678
 
+  let host = Host.create('linux')
   let osKind = os.Kind.for('omnios')
   let architecture = arch.Architecture.for(
     arch.Kind.x86_64,
@@ -18,7 +19,7 @@ describe('OmniOS QemuVm', () => {
     osKind,
     host.hypervisor
   )
-  let input = new Input()
+  let input = new Input(host)
   let config = {
     memory: memory,
     cpuCount: cpuCount,
@@ -26,9 +27,7 @@ describe('OmniOS QemuVm', () => {
     ssHostPort: ssHostPort,
     cpu: '',
     machineType: '',
-    uuid: '',
     resourcesDiskImage: '',
-    userboot: '',
     firmware: ''
   }