+++ src/hypervisor.ts
@@ -1,20 +1,16 @@
 import {Architecture} from './architecture'
-import {ResourceUrls} from './operating_systems/resource_urls'
 import {Vm as QemuVm} from './qemu_vm'
-import {Vm as XhyveVm} from './xhyve_vm'
 import {getOrDefaultOrThrow} from './utility'
 
 export enum Kind {
-  xhyve,
   qemu
 }
 
 export function toKind(value: string): Kind | undefined {
-  return architectureMap[value.toLocaleLowerCase()]
+  return kindMap[value.toLocaleLowerCase()]
 }
 
-const architectureMap: Record<string, Kind> = {
-  xhyve: Kind.xhyve,
+const kindMap: Record<string, Kind> = {
   qemu: Kind.qemu
 } as const
 
@@ -22,41 +18,10 @@ export interface Hypervisor {
   get kind(): Kind
   get sshPort(): number
   get firmwareFile(): string
-  get vmModule(): typeof QemuVm | typeof XhyveVm
+  get vmModule(): typeof QemuVm
   get efi(): Hypervisor
   getResourceUrl(architecture: Architecture): string
   resolve<T>(implementation: Record<string, T>): T
-}
-
-export class Xhyve implements Hypervisor {
-  get kind(): Kind {
-    return Kind.xhyve
-  }
-
-  get sshPort(): number {
-    return 22
-  }
-
-  get firmwareFile(): string {
-    return 'uefi.fd'
-  }
-
-  get vmModule(): typeof XhyveVm {
-    return XhyveVm
-  }
-
-  get efi(): Hypervisor {
-    return this
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  getResourceUrl(_architecture: Architecture): string {
-    return `${ResourceUrls.create().resourceBaseUrl}/xhyve-macos.tar`
-  }
-
-  resolve<T>(implementation: Record<string, T>): T {
-    return getOrDefaultOrThrow(implementation, 'xhyve')
-  }
 }
 
 export class Qemu implements Hypervisor {
@@ -97,11 +62,10 @@ export class QemuEfi extends Qemu {
   }
 }
 
-export function toHypervisor(kind: Kind): typeof Xhyve | typeof Qemu {
+export function toHypervisor(kind: Kind): typeof Qemu {
   return hypervisorMap[kind]
 }
 
-const hypervisorMap: Record<Kind, typeof Xhyve | typeof Qemu> = {
-  [Kind.xhyve]: Xhyve,
+const hypervisorMap: Record<Kind, typeof Qemu> = {
   [Kind.qemu]: Qemu
 } as const